director-aiohttp/requirements.txt
aiohttp
aiohttp-swagger
aiohttp-devtools
python-socketio
requests